+++ src/FPIndia/Model.hs
@@ -1,0 +1,8 @@
+module FPIndia.Model where
+
+import Ema.Route.Lib.Extra.StaticRoute qualified as SR
+
+data Model = Model
+  { modelStatic :: SR.Model
+  }
+  deriving stock (Eq, Show, Generic)
